initiate db client
