+++ index.js
@@ -1,1 +1,2 @@
+import './services/env.js'
 import './app/index.js'
