+++ services/env.js
@@ -1,0 +1,3 @@
+import * as dotenv from 'dotenv'
+
+dotenv.config()
